Hexa-004 - User class changes

+++ backend/src/main/java/com/hexaware/hexacliq/jwt/JwtUserDetails.java
@@ -15,12 +15,12 @@ public class JwtUserDetails implements UserDetails {
 
 	private static final long serialVersionUID = 5155720064139820502L;
 
-	private final Integer id;
+	private final Long id;
 	private final String username;
 	private final String password;
 	private final Collection<? extends GrantedAuthority> authorities;
 
-	public JwtUserDetails(Integer id, String username, String password, String role) {
+	public JwtUserDetails(Long id, String username, String password, String role) {
 		this.id = id;
 		this.username = username;
 		this.password = password;
